refactored tests to reuse refactored common goa data file methods
