+++ indexing/src/test/java/uk/ac/ebi/quickgo/index/common/datafile/GOADataFileParsingHelperTest.java
@@ -1,0 +1,110 @@
+package uk.ac.ebi.quickgo.index.common.datafile;
+
+import java.util.Arrays;
+import java.util.Map;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasEntry;
+import static org.hamcrest.Matchers.isEmptyString;
+import static org.hamcrest.collection.IsMapContaining.hasKey;
+import static org.hamcrest.core.Is.is;
+import static uk.ac.ebi.quickgo.index.common.datafile.GOADataFileParsingHelper.convertLinePropertiesToMap;
+import static uk.ac.ebi.quickgo.index.common.datafile.GOADataFileParsingUtil.concatProperty;
+import static uk.ac.ebi.quickgo.index.common.datafile.GOADataFileParsingUtil.concatStrings;
+
+/**
+ * Created 28/04/16
+ * @author Edd
+ */
+public class GOADataFileParsingHelperTest {
+    private static final String INTER_VALUE_DELIMITER = "|";
+    private static final String INTER_VALUE_DELIMITER_REGEX = "\\|";
+    public static final String INTRA_VALUE_DELIMITER = "=";
+
+    @Rule
+    public ExpectedException thrown = ExpectedException.none();
+
+    @Test
+    public void nullInterValueDelimiterThrowsException() throws Exception {
+        String propsText = "";
+
+        thrown.expect(AssertionError.class);
+        thrown.expectMessage("InterValueDelimiter cannot be null");
+
+        convertLinePropertiesToMap(propsText, null, INTRA_VALUE_DELIMITER);
+    }
+
+    @Test
+    public void nullIntraValueDelimiterThrowsException() throws Exception {
+        String propsText = "";
+
+        thrown.expect(AssertionError.class);
+        thrown.expectMessage("IntraValueDelimiter cannot be null");
+
+        convertLinePropertiesToMap(propsText, INTER_VALUE_DELIMITER, null);
+    }
+
+    @Test
+    public void nullPropertiesTextReturnsEmptyMap() throws Exception {
+        Map<String, String>
+                propsMap = convertLinePropertiesToMap(null, INTER_VALUE_DELIMITER_REGEX, INTRA_VALUE_DELIMITER);
+
+        assertThat(propsMap.isEmpty(), is(true));
+    }
+
+    @Test
+    public void emptyPropertiesTextReturnsEmptyMap() throws Exception {
+        String propsText = "";
+
+        Map<String, String> propsMap = convertLinePropertiesToMap(propsText, INTER_VALUE_DELIMITER_REGEX, INTRA_VALUE_DELIMITER);
+
+        assertThat(propsMap.isEmpty(), is(true));
+    }
+
+    @Test
+    public void singlePropertyWithNoValueReturnsMapWithSingleEntryWithKeyAndNoValue() throws Exception {
+        String propKey = "key";
+        String propsText = propKey;
+
+        Map<String, String> propsMap = convertLinePropertiesToMap(propsText, INTER_VALUE_DELIMITER_REGEX, INTRA_VALUE_DELIMITER);
+
+        assertThat(propsMap.size(), is(1));
+        assertThat(propsMap, hasKey(propKey));
+        assertThat(propsMap.get(propKey), isEmptyString());
+    }
+
+    @Test
+    public void singlePropertyWithValueReturnsMapWithSingleEntry() throws Exception {
+        String propKey = "key";
+        String propValue = "value";
+
+        String propsText = concatProperty(propKey, propValue, INTRA_VALUE_DELIMITER);
+
+        Map<String, String> propsMap = convertLinePropertiesToMap(propsText, INTER_VALUE_DELIMITER_REGEX, INTRA_VALUE_DELIMITER);
+
+        assertThat(propsMap.size(), is(1));
+        assertThat(propsMap, hasEntry(propKey, propValue));
+    }
+
+    @Test
+    public void twoPropertiesReturnMapWithTwoEntries() throws Exception {
+        String propKey1 = "key1";
+        String propValue1 = "value1";
+        String concatProp1 = concatProperty(propKey1, propValue1, INTRA_VALUE_DELIMITER);
+
+        String propKey2 = "key2";
+        String propValue2 = "value2";
+        String concatProp2 = concatProperty(propKey2, propValue2, INTRA_VALUE_DELIMITER);
+
+        String propsText = concatStrings(Arrays.asList(concatProp1, concatProp2), INTER_VALUE_DELIMITER);
+
+        Map<String, String> propsMap = convertLinePropertiesToMap(propsText, INTER_VALUE_DELIMITER_REGEX, INTRA_VALUE_DELIMITER);
+
+        assertThat(propsMap.size(), is(2));
+        assertThat(propsMap, hasEntry(propKey1, propValue1));
+        assertThat(propsMap, hasEntry(propKey2, propValue2));
+    }
+}
