+++ indexing/src/test/java/uk/ac/ebi/quickgo/index/geneproduct/GeneProductUtil.java
@@ -1,10 +1,7 @@
 package uk.ac.ebi.quickgo.index.geneproduct;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 /**
- * Helper class that houses several methods used in various tests classes.
+ * Helper class to house methods used in various tests classes.
  *
  * @author Ricardo Antunes
  */
@@ -15,11 +12,4 @@ class GeneProductUtil {
         return "taxon:" + taxonId;
     }
 
-    static String concatStrings(List<String> values, String delimiter) {
-        return values.stream().collect(Collectors.joining(delimiter));
-    }
-
-    static String concatProperty(String key, String value, String delimiter) {
-        return key + delimiter + value;
-    }
 }
